requirements.txt
aiodocker==0.21.0
aiohttp-pydantic==1.12.2
aiohttp==3.8.5
asyncssh==2.13.1
pydantic==1.10.7
PyYAML==6.0.1
trafaret==2.1.1
pyjwt==2.8.0